Refactor recast navigation and split jump/fall into extensions
